Crate tinycdb

Rust bindings to the TinyCDB library.

TinyCDB is a very fast and simple package for creating and reading constant databases, as introduced by Dan Bernstein in his cdb package.

CDB is a constant database, that is, it cannot be updated at a runtime, only rebuilt. Rebuilding is atomic operation and is very fast - much faster than of many other similar packages. Once created, CDB may be queried, and a query takes very little time to complete.

Structs

Cdb
The Cdb struct represents an open instance of a CDB database.

CdbCreator
The CdbCreator struct is used while building a new CDB instance.

CdbError
Our error type

CdbIterator
A CdbIterator allows iterating over all the keys in a CDB database.

Enums

CdbErrorKind
Kinds of errors that can be encountered.

CdbPutMode
CdbPutMode represents the different behaviours that will be used when inserting a key into a database where the key already exists.

Type Definitions

CdbResult
A specialized Result type that might contain a CdbError.